Promote form_with subkit to form kit
