+++ playbook/app/pb_kits/playbook/pb_form/docs/_form_form_with_validate.html.erb
@@ -13,8 +13,8 @@
   ]
 %>
 
-<%= pb_rails("form/form_with", props: {
-    scope: :example, method: :get, url: "",
+<%= pb_rails("form", props: {
+    options: { scope: :example, method: :get, url: "" },
     validate: true
   }) do |form| %>
   <%= form.text_field :example_text_field, props: { label: true, required: true } %>
