+++ playbook/app/pb_kits/playbook/pb_form/form.rb
@@ -3,33 +3,41 @@
 module Playbook
   module PbForm
     class Form < ::Playbook::KitBase
-      prop :form_system, type: Playbook::Props::Enum,
-                         values: %w[form_with],
-                         default: "form_with"
-      prop :form_system_options, type: Playbook::Props::Base
+      prop :form_system, deprecated: "Playbook only supports form_with",
+                         type: Playbook::Props::Base
+      prop :form_system_options, deprecated: "Use options instead",
+                                 type: Playbook::Props::Base
+      prop :options, type: Playbook::Props::Base
       prop :validate, type: Playbook::Props::Boolean, default: false
 
-      def form_system_options
-        form_options = prop(:form_system_options)
-        if form_options.is_a?(::Array)
-          model, options = *form_options
-          (options || {}).merge(model: model)
-        else
-          form_options
-        end
+      def data
+        prop(:data).merge("pb-form-validation" => validate)
       end
 
-      def render_in(view_context, &block)
-        form_kit.render_in(view_context, &block)
+      def form_options
+        {
+          id: id,
+          aria: aria,
+          class: classname,
+          data: data,
+          builder: ::Playbook::PbForm::FormBuilder,
+        }.merge(options)
       end
 
-    private
+      def options
+        prop(:options) || prop(:form_system_options) || { }
+      end
 
-      def form_kit
-        @form_system ||= begin
-          ::Playbook::KitResolver.resolve("form/#{form_system}")
-                                 .new(options: form_system_options, validate: validate, children: children)
-        end
+      def render_form(builder)
+        view_context.capture(builder, &children)
+      end
+
+      def classname
+        [generate_classname("pb-form"), options[:class]].join(" ")
+      end
+
+      def render_in(view_context, &_block)
+        super(view_context, &nil)
       end
     end
   end
