+++ playbook/app/pb_kits/playbook/pb_form/form_builder.rb
@@ -2,24 +2,20 @@
 
 module Playbook
   module PbForm
-    module FormBuilder
-      extend ActiveSupport::Concern
-
-      included do
-        prepend(FormFieldBuilder.new(:email_field, kit_name: "text_input"))
-        prepend(FormFieldBuilder.new(:number_field, kit_name: "text_input"))
-        prepend(FormFieldBuilder.new(:search_field, kit_name: "text_input"))
-        prepend(FormFieldBuilder.new(:telephone_field, kit_name: "text_input"))
-        prepend(FormFieldBuilder.new(:text_field, kit_name: "text_input"))
-        prepend(FormFieldBuilder.new(:password_field, kit_name: "text_input"))
-        prepend(FormFieldBuilder.new(:url_field, kit_name: "text_input"))
-        prepend(FormFieldBuilder.new(:text_area, kit_name: "textarea"))
-        prepend(TypeaheadField)
-        prepend(SelectField)
-        prepend(CollectionSelectField)
-        prepend(CheckboxField)
-        prepend(DatePickerField)
-      end
+    class FormBuilder < ::ActionView::Helpers::FormBuilder
+      prepend(FormFieldBuilder.new(:email_field, kit_name: "text_input"))
+      prepend(FormFieldBuilder.new(:number_field, kit_name: "text_input"))
+      prepend(FormFieldBuilder.new(:search_field, kit_name: "text_input"))
+      prepend(FormFieldBuilder.new(:telephone_field, kit_name: "text_input"))
+      prepend(FormFieldBuilder.new(:text_field, kit_name: "text_input"))
+      prepend(FormFieldBuilder.new(:password_field, kit_name: "text_input"))
+      prepend(FormFieldBuilder.new(:url_field, kit_name: "text_input"))
+      prepend(FormFieldBuilder.new(:text_area, kit_name: "textarea"))
+      prepend(TypeaheadField)
+      prepend(SelectField)
+      prepend(CollectionSelectField)
+      prepend(CheckboxField)
+      prepend(DatePickerField)
 
       def actions
         @template.send(:view_context).content_tag :ol, class: "pb-form-actions" do
