+++ playbook/app/pb_kits/playbook/pb_form/form_builder/typeahead_field.rb
@@ -2,7 +2,7 @@
 
 module Playbook
   module PbForm
-    module FormBuilder
+    class FormBuilder
       module TypeaheadField
         def typeahead(name, _options = {}, _html_options = {}, props: {})
           props[:name] = name
